BlogAPI/requirements.txt
djangorestframework
psycopg2-binary
python-decouple
Pillow
